arch/riscv: Store mprv bit in size_t
CSRs are XLEN bits wide (i.e. the same width as general purpose
registers), so size_t seems a little more correct than int.

This change doesn't affect functionality because MSTATUS_MPRV already
fits in 31 bits.

Reviewed-on: https://review.coreboot.org/25625
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/arch/riscv/include/vm.h
@@ -42,7 +42,7 @@ void mstatus_init(void); // need to setup mstatus so we know we have virtual mem
 	static inline type name(type *p);				\
 	static inline type name(type *p)				\
 	{								\
-		int mprv = MSTATUS_MPRV;				\
+		size_t mprv = MSTATUS_MPRV;				\
 		type value;						\
 		asm (							\
 			"csrs		mstatus, %1\n"			\
@@ -57,7 +57,7 @@ void mstatus_init(void); // need to setup mstatus so we know we have virtual mem
 	static inline void name(type *p, type value);			\
 	static inline void name(type *p, type value)			\
 	{								\
-		int mprv = MSTATUS_MPRV;				\
+		size_t mprv = MSTATUS_MPRV;				\
 		asm (							\
 			"csrs		mstatus, %0\n"			\
 			STRINGIFY(insn) " %1, 0(%2)\n"			\
